+++ weather_app/ui/static/styles.css
